added:  emC_srcApplSpec/applConv/fatalError_printf_exitMain.c
* The emC_srcApplSpec/SimpleNumCNoExc/ApplSimpleStop_emC.c designated as "do not use"
TODO the fatal error abort behaviour should be better documented and cleaned.
* rename: StringBase.*: set_s0len_StringJc from set_StringJc
* rename: StingJc.h, StringBufferJc.c: set_s_StringJc from set_StringJc
  both identifier were conflicting after enhancements, should be better clarified.
*Docu enhanced, some small changes

+++ emC/Jc/ReflMemAccessJc.c
@@ -539,7 +539,7 @@ int32 accessTarget_Inspc(Cmd_InspcTargetProxy_e cmd, int device, uint32 address,
 
 /**Gets any info from a remoteCPU, which has max. 8 byte return value. */
 #ifndef DEF_INSPC_REMOTEACCESS
-int64 accessTarget64_Inspc(Cmd_InspcTargetProxy_e cmd, int device, uint32 address, int64 input)
+int64 accessTarget64_Inspc(Cmd_InspcTargetProxy_e cmd, int device, uint64 address, int64 input)
 {
 	return 0;
 }
